refactor: change layout to page in file names
