+++ package.json
@@ -31,7 +31,8 @@
     "react-dom": "^19.0.0",
     "react-router": "^7.5.3",
     "tailwind-merge": "^3.2.0",
-    "tailwindcss": "^4.1.5"
+    "tailwindcss": "^4.1.5",
+    "valibot": "^1.1.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.8.0",
